loadoption: Make things safer
Do not store the unsafe cast pointer to the byte array in the
LoadOption struct, but store a slice, and convert to C array
on demand.

Also pass the size of the data to efi_loadopt_desc() as we
probably should be doing; it's passed onward to the UCS-2
to UTF-8 conversion code.