efibootmgr: Add function to maybe update file
This will be used to copy kernels over if they are different
(avoiding unnecessary rewrites).

This adds the afero dependency for the test implementation, we
should work on removing the need for it, as it blows up go.sum,
but at least it does not blow up the binary size due to only
being imported in test files.