Trust boot assets used in the current boot
The resealing code computes PCR digests from assets unpacked to the ESP,
which depends on a list of hashes of boot assets that have previously
been established as trusted by nullboot. This list of hashes doesn't
exist on first boot though - the trust here is established by the
provisioning of the image, which trusts the initial boot assets. To handle
this case, always trust the boot assets used in the current boot,
determined by inspecting the TCG log.