nullbootctl: Sync the ESP and improve comments
Sync the ESP between copying in the kernels and adding
them to the boot order so they are written to the backing
storage before the boot order is updated.

This also ensures that when kernels are removed later,
the new kernels are there so there always is a bootable
path.

Also add additional documentation strings.