feat: Determine the state of the bootloader
To follow the introduction of the fallback mechanism, this change allows
for a query on the state of the bootloader. i.e. if the fallback
mechanism is enabled and a new kernel boots, the BootOrder EFI variable
and the rest of the bootloader can be amended. Otherwise, if the
BootCurrent EFI variable is the first boot entry in the BootOrder, no
changes are needed.