feat: CLI update to support fallback mechanism
This change enables the kernel fallback mechanism as it relates to new
kernels via `nullbootctl`. The new CLI flag disables the setting of
BootNext such that an unsuccessful boot onto BootNext does not result
in the continued setting of BootNext. Rather, the installation trigger
will set BootNext and the systemd service executed on startup will not.

Default behavior is to implement the fallback mechanism, and a successful
boot onto that kernel will amend the boot loader. Unsuccessful boots
will not amend the boot loader or set BootNext again until another apt
event triggers `nullbootctl` again.