feat: Systemd one-shot service for Kernel fallback mechanism
Adds a systemd one-shot service to amend the bootloader on a successfully
booted, new kernel entry using the -no-boot-next option.

Should a newly installed kernel panic while booting, no changes will
occur in the boot loader and the service ensures BootNext is not set
again on the bad kernel (unless another evocation of nullbootctl does).

In the event that a new kernel successfully boots, nullboot will amend
the boot loader with the new kernel.